+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Creación de la entidad ``Usuario`` (Alberto Jiménez Expósito)
 - Creación del servicio ``ServicioIncidencias`` (María Ximena Galdames Fernandes)
 - Creación de la entidad ``Incidencia`` (Carlos Mayor Navarro)
+- Añadido atributo `loginUsuario` a ``Incidencia`` para indicar el `Usuario` que la registró (Alberto Jiménez Expósito)
+- Añadido método ``nuevaIncidencia`` que registra una nueva incidencia en el sistema (Alberto Jiménez Expósito)